pyproject.toml
[tool.poetry]
name = "research-paper-implementations"
version = "0.1.0"
description = ""
authors = ["Archit <architojha420@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jax = "^0.5.0"
flax = "^0.10.2"
matplotlib = "^3.10.0"
rich = "^13.9.4"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
optax = "^0.2.4"
tqdm = "^4.67.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
